refactor(UserTable): move buildFismaSystemsMap to its own module to fix lint warnings.
